+++ test_app/spec-opal/example_spec.js.rb
@@ -1,3 +1,5 @@
+require 'spec_helper'
+
 describe 'a running spec' do
   it 'runs successfully' do
     'I run'.should =~ /run/