+++ test_app/spec-opal/spec_helper.rb
@@ -1,0 +1,3 @@
+require 'opal'
+require 'opal-rspec'
+Opal::RSpec::Runner.autorun