+++ test_app/spec-opal/subdirectory/other_spec.js.rb
@@ -1,3 +1,5 @@
+require 'spec_helper'
+
 describe "spec inside subdirectory" do
   it "runs fine" do
     1.should == 1